refactor: updated directory structure of providers
